+++ example/annotations/annotations.dart
@@ -7,12 +7,13 @@ main() {
   var editor = ace.edit(querySelector('#editor'))
   ..theme = new ace.Theme('ace/theme/monokai')
   ..session.mode = new ace.Mode('ace/mode/dart')
-  ..session.annotations = [
+  ..session.setAnnotations([
     const ace.Annotation(
       row: 1,
       text: 'hello'),
     const ace.Annotation(
       row: 3,
       html: '<span><i>ruh-roh</i></span>', 
-      type: ace.Annotation.ERROR) ];
+      type: ace.Annotation.ERROR)
+  ]);
 }