+++ test/src/edit_session_test.dart
@@ -412,50 +412,53 @@ void testRemoveGutterDecoration() {
 
 @Test()
 void testGetAnnotations() {
-  expect(session.annotations, isEmpty); 
+  expect(session.getAnnotations(), isEmpty); 
 }
 
 @Test()
 void testSetAnnotations() {
   session.onChangeAnnotation.listen(expectAsync1((_) {
-    expect(session.annotations.length, equals(2));
-    expect(session.annotations[0], equals(const Annotation(
+    final annotations = session.getAnnotations();
+    expect(annotations.length, equals(2));
+    expect(annotations[0], equals(const Annotation(
       row: 42,
       text: 'snarf')));
-    expect(session.annotations[1], equals(const Annotation(
+    expect(annotations[1], equals(const Annotation(
       row: 16,
       html: '<span>ruh-roh</span>', 
       type: Annotation.ERROR)));
   }));
-  session.annotations = [
+  session.setAnnotations([
     const Annotation(row: 42, text: 'snarf'),
     const Annotation(
       row: 16,
       html: '<span>ruh-roh</span>', 
-      type: Annotation.ERROR) ];
+      type: Annotation.ERROR) 
+  ]);
 }
 
 @Test()
 void testClearAnnotations() {
-  session.annotations = [const Annotation(row: 42, text: 'foo')];
-  expect(session.annotations.length, equals(1));
-  expect(session.annotations[0], equals(
+  session.setAnnotations([const Annotation(row: 42, text: 'foo')]);
+  final annotations = session.getAnnotations();
+  expect(annotations.length, equals(1));
+  expect(annotations[0], equals(
     const Annotation(row: 42, text: 'foo')));
   session.onChangeAnnotation.listen(expectAsync1((_) {
-    expect(session.annotations, isEmpty);
+    expect(session.getAnnotations(), isEmpty);
   }));
   session.clearAnnotations();
 }
 
 @Test() 
 void testGetBreakpoints() {
-  expect(session.breakpoints, isEmpty);
+  expect(session.getBreakpoints(), isEmpty);
 }
 
 @Test()
 void testSetBreakpoint() {
   session.onChangeBreakpoint.listen(expectAsync1((_) {
-    expect(session.breakpoints[2], equals('ace_breakpoint'));
+    expect(session.getBreakpoints()[2], equals('ace_breakpoint'));
   }));  
   session.setBreakpoint(2);
 }
@@ -463,9 +466,10 @@ void testSetBreakpoint() {
 @Test()
 void testSetBreakpoints() {
   session.onChangeBreakpoint.listen(expectAsync1((_) {
-    expect(session.breakpoints[2], isNotNull);
-    expect(session.breakpoints[3], isNotNull);
-    expect(session.breakpoints[4], isNotNull);
+    final breakpoints = session.getBreakpoints();
+    expect(breakpoints[2], isNotNull);
+    expect(breakpoints[3], isNotNull);
+    expect(breakpoints[4], isNotNull);
   }));  
   session.setBreakpoints([2, 3, 4]);
 }
@@ -473,7 +477,7 @@ void testSetBreakpoints() {
 @Test()
 void testSetBreakpointWithClassName() {
   session.onChangeBreakpoint.listen(expectAsync1((_) {
-    expect(session.breakpoints[4], equals('fancy_breakpoint'));
+    expect(session.getBreakpoints()[4], equals('fancy_breakpoint'));
   }));  
   session.setBreakpoint(4, className: 'fancy_breakpoint');
 }
@@ -481,9 +485,9 @@ void testSetBreakpointWithClassName() {
 @Test()
 void testClearBreakpoint() {
   session.setBreakpoint(3);
-  expect(session.breakpoints[3], isNotNull);
+  expect(session.getBreakpoints()[3], isNotNull);
   session.onChangeBreakpoint.listen(expectAsync1((_) {
-    expect(session.breakpoints[3], isNull);
+    expect(session.getBreakpoints()[3], isNull);
   }));
   session.clearBreakpoint(3);
 }
@@ -491,10 +495,10 @@ void testClearBreakpoint() {
 @Test()
 void testClearBreakpoints() {
   session.setBreakpoints([1, 4]);
-  expect(session.breakpoints[1], isNotNull);
-  expect(session.breakpoints[4], isNotNull);
+  expect(session.getBreakpoints()[1], isNotNull);
+  expect(session.getBreakpoints()[4], isNotNull);
   session.onChangeBreakpoint.listen(expectAsync1((_) {
-    expect(session.breakpoints, isEmpty);
+    expect(session.getBreakpoints(), isEmpty);
   }));
   session.clearBreakpoints();
 }