+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 # Ace.dart Changes
 
-## 0.1.8-dev+1.8.2014
+## 0.2.0-dev+1.8.2014
 
+- Replaced the `EditSession.breakpoints` getter with the `getBreakpoints` 
+method.
+- Split `EditSession.annotations` getter/setter into the `getAnnotations` and
+`setAnnotations` methods; this matches `ace.js` and helps to clarify their 
+separate behavior.
 - Added the `new Mode.named` factory, `name` getter, and enumeration constants 
 to `Mode`.
 - Added the `onChangeSelection` event stream to `Editor`.